Update test-geneator.py
The procedure is updated such that generates single IPs for source and destinations. A random variable decides if repeat the latest traffic, pick one of the previous traffics, or generated a new traffic. The probability for picking from history is 30%, for repeating the latest one is 50%, and for new traffic is 20%. Minor error fixed.